Added new zodiac.vcproj and casts to stop VS2005 warnings.

+++ trunk/zkdecrypto/headers/solve.h
@@ -343,10 +343,10 @@ void SetLanguage()
 
 	switch(iLang)
 	{
-		case 0: strcpy(szLanguage,"English"); strcpy(szLang,LANG_ENG); fLangIoC=IOC_ENG; break;
-		case 1: strcpy(szLanguage,"Spanish"); strcpy(szLang,LANG_SPA); fLangIoC=IOC_SPA; break;
-		case 2: strcpy(szLanguage,"German"); strcpy(szLang,LANG_GER); fLangIoC=IOC_GER; break;
-		case 3: strcpy(szLanguage,"Italian"); strcpy(szLang,LANG_ITA); fLangIoC=IOC_ITA; break;
+		case 0: strcpy(szLanguage,"English"); strcpy(szLang,LANG_ENG); fLangIoC=(float)IOC_ENG; break;
+		case 1: strcpy(szLanguage,"Spanish"); strcpy(szLang,LANG_SPA); fLangIoC=(float)IOC_SPA; break;
+		case 2: strcpy(szLanguage,"German"); strcpy(szLang,LANG_GER); fLangIoC=(float)IOC_GER; break;
+		case 3: strcpy(szLanguage,"Italian"); strcpy(szLang,LANG_ITA); fLangIoC=(float)IOC_ITA; break;
 	}
 	
 	SetIoC(fLangIoC);
